[wip] scummvm midi adlib driver
